+++ pennylane/_device.py
@@ -39,7 +39,8 @@ user interface:
 .. autosummary::
     short_name
     capabilities
-    supported
+    supports_operation
+    supports_observable
     execute
     reset
 
@@ -89,6 +90,7 @@ Code details
 import abc
 
 import autograd.numpy as np
+from pennylane.operation import Operation, Observable
 
 
 class DeviceError(Exception):
@@ -182,7 +184,7 @@ class Device(abc.ABC):
         return cls._capabilities
 
     def execute(self, queue, observables):
-        """Execute a queue of quantum operations on the device and then measure the given expectation values.
+        """Execute a queue of quantum operations on the device and then measure the given observables.
 
         For plugin developers: Instead of overwriting this, consider implementing a suitable subset of
         :meth:`pre_apply`, :meth:`apply`, :meth:`post_apply`, :meth:`pre_measure`,
@@ -216,13 +218,25 @@ class Device(abc.ABC):
                     results.append(self.expval(obs.name, obs.wires, obs.parameters))
                 elif obs.return_type == "variance":
                     results.append(self.var(obs.name, obs.wires, obs.parameters))
+                elif obs.return_type == "sample":
+                    if not hasattr(obs, "num_samples"):
+                        raise DeviceError("Number of samples not specified for observable {}".format(obs.name))
+
+                    results.append(np.array(self.sample(obs.name, obs.wires, obs.parameters, obs.num_samples)))
 
             self.post_measure()
 
             self._op_queue = None
             self._obs_queue = None
 
-            return np.array(results)
+            # Ensures that a combination with sample does not put
+            # expvals and vars in superfluous arrays
+            if all(obs.return_type == "sample" for obs in observables):
+                return np.asarray(results)
+            if any(obs.return_type == "sample" for obs in observables):
+                return np.asarray(results, dtype="object")
+
+            return np.asarray(results)
 
     @property
     def op_queue(self):
@@ -289,16 +303,38 @@ class Device(abc.ABC):
 
         return MockContext()
 
-    def supported(self, name):
-        """Checks if an operation or observable is supported by this device.
+    def supports_operation(self, operation):
+        """Checks if an operation is supported by this device.
 
         Args:
-            name (str): name of the operation or observable
+            operation (Operation,str): operation to be checked
 
         Returns:
-            bool: True iff it is supported
+            bool: ``True`` iff supplied operation is supported
         """
-        return name in self.operations.union(self.observables)
+        if isinstance(operation, type) and issubclass(operation, Operation):
+            return operation.__name__ in self.operations
+        if isinstance(operation, str):
+            return operation in self.operations
+
+        raise ValueError("The given operation must either be a pennylane.Operation class or a string.")
+
+    def supports_observable(self, observable):
+        """Checks if an observable is supported by this device.
+
+        Args:
+            operation (Observable,str): observable to be checked
+
+        Returns:
+            bool: ``True`` iff supplied observable is supported
+        """
+        if isinstance(observable, type) and issubclass(observable, Observable):
+            return observable.__name__ in self.observables
+        if isinstance(observable, str):
+            return observable in self.observables
+
+        raise ValueError("The given operation must either be a pennylane.Observable class or a string.")
+
 
     def check_validity(self, queue, observables):
         """Checks whether the operations and observables in queue are all supported by the device.
@@ -353,9 +389,27 @@ class Device(abc.ABC):
           par (tuple[float]): parameter values
 
         Returns:
-          float: variance :math:`\mathrm{var}(A) = \bra{\psi}A^2\ket{\psi} - \bra{\psi}A\ket{\psi}^2`
-            """
-        raise NotImplementedError
+            float: variance :math:`\mathrm{var}(A) = \bra{\psi}A^2\ket{\psi} - \bra{\psi}A\ket{\psi}^2`
+        """
+        raise NotImplementedError("Returning variances from QNodes not currently supported by {}".format(self.short_name))
+
+    def sample(self, observable, wires, par, n=None):
+        """Return a sample of an observable.
+
+        For plugin developers: this function should return the result of an evaluation
+        of the given observable on the device.
+
+        Args:
+            observable (str): name of the observable
+            wires (Sequence[int]): subsystems the observable is to be measured on
+            par (tuple): parameters for the observable
+            n (int): Number of samples that should be obtained. Defaults to the
+                number of shots given as a parameter to the corresponding Device.
+
+        Returns:
+            array[float]: samples in an array of dimension ``(n, num_wires)``
+        """
+        raise NotImplementedError("Returning samples from QNodes not currently supported by {}".format(self.short_name))
 
     @abc.abstractmethod
     def reset(self):